+++ python/pathway/xpacks/llm/parsers.py
@@ -14,7 +14,7 @@ import warnings
 from collections.abc import Callable
 from functools import partial
 from io import BytesIO
-from typing import TYPE_CHECKING, Any, Literal
+from typing import Any, Literal
 
 from PIL import Image
 from pydantic import BaseModel
@@ -25,9 +25,6 @@ from pathway.internals.config import _check_entitlements
 from pathway.optional_import import optional_imports
 from pathway.xpacks.llm import _parser_utils, llms, prompts
 from pathway.xpacks.llm.constants import DEFAULT_VISION_MODEL
-
-if TYPE_CHECKING:
-    from openparse.processing import IngestionPipeline
 
 logger = logging.getLogger(__name__)
 
@@ -350,167 +347,6 @@ class DoclingParser(pw.UDF):
 
     async def __wrapped__(self, contents: bytes, **kwargs) -> list[tuple[str, dict]]:
         return await self.parse(contents)
-
-
-class OpenParse(pw.UDF):
-    """
-    Parse PDFs using `open-parse library <https://github.com/Filimoa/open-parse>`_.
-
-    When used in the
-    `VectorStoreServer <https://pathway.com/developers/api-docs/pathway-xpacks-llm/
-    vectorstore#pathway.xpacks.llm.vector_store.VectorStoreServer>`_,
-    splitter can be set to ``None`` as OpenParse already chunks the documents.
-
-    Args:
-        table_args: dictionary containing the table parser arguments. Needs to have key ``parsing_algorithm``,
-            with the value being one of ``"llm"``, ``"unitable"``, ``"pymupdf"``, ``"table-transformers"``.
-            ``"llm"`` parameter can be specified to modify the vision LLM used for parsing.
-            Will default to ``OpenAI`` ``gpt-4o``, with markdown table parsing prompt.
-            Default config requires ``OPENAI_API_KEY`` environment variable to be set.
-            For information on other parsing algorithms and supported arguments check
-            `the OpenParse documentation <https://filimoa.github.io/open-parse/processing/parsing-tables/overview/>`_.
-        image_args: dictionary containing the image parser arguments.
-            Needs to have the following keys ``parsing_algorithm``, ``llm``, ``prompt``.
-            Currently, only supported ``parsing_algorithm`` is ``"llm"``.
-            ``"llm"`` parameter can be specified to modify the vision LLM used for parsing.
-            Will default to ``OpenAI`` ``gpt-4o``, with markdown image parsing prompt.
-            Default config requires ``OPENAI_API_KEY`` environment variable to be set.
-        parse_images: whether to parse the images from the PDF. Detected images will be
-            indexed by their description from the parsing algorithm.
-            Note that images are parsed with separate OCR model, parsing may take a while.
-        processing_pipeline: str or IngestionPipeline.
-            Specifies the pipeline used for post-processing extracted elements.
-            - ``"pathway_pdf_default"``: Uses ``SimpleIngestionPipeline`` from Pathway.
-            This is a simple processor that combines close elements, combines the headers
-            with the text body, and removes weirdly formatted/small elements.
-            Can be set with: ``"pathway_pdf_default"`` or using the class,
-            ``from pathway.xpacks.llm.openparse_utils import SimpleIngestionPipeline``.
-            - ``"merge_same_page"``: Uses ``SamePageIngestionPipeline`` to chunk based on pages.
-            - Any other pipeline from the ``openparse.processing`` can also be used.
-            Defaults to ``SimpleIngestionPipeline``.
-        cache_strategy: Defines the caching mechanism. To enable caching,
-            a valid :py:class:``~pathway.udfs.CacheStrategy`` should be provided.
-            Defaults to None.
-
-    Example:
-
-    >>> import pathway as pw
-    >>> from pathway.xpacks.llm import llms, parsers, prompts
-    >>> chat = llms.OpenAIChat(model="gpt-4o")
-    >>> table_args = {
-    ...    "parsing_algorithm": "llm",
-    ...    "llm": chat,
-    ...    "prompt": prompts.DEFAULT_MD_TABLE_PARSE_PROMPT,
-    ... }
-    >>> image_args = {
-    ...     "parsing_algorithm": "llm",
-    ...     "llm": chat,
-    ...     "prompt": prompts.DEFAULT_IMAGE_PARSE_PROMPT,
-    ... }
-    >>> parser = parsers.OpenParse(table_args=table_args, image_args=image_args)
-    """
-
-    def __init__(
-        self,
-        table_args: dict | None = None,
-        image_args: dict | None = None,
-        parse_images: bool = False,
-        processing_pipeline: IngestionPipeline | str | None = None,
-        cache_strategy: udfs.CacheStrategy | None = None,
-    ):
-        with optional_imports("xpack-llm-docs"):
-            import openparse  # noqa:F401
-            from pypdf import PdfReader  # noqa:F401
-
-            from .openparse_utils import (
-                PyMuDocumentParser,
-                SamePageIngestionPipeline,
-                SimpleIngestionPipeline,
-            )
-
-        super().__init__(cache_strategy=cache_strategy)
-
-        if table_args is None:
-            table_args = {
-                "parsing_algorithm": "llm",
-                "llm": DEFAULT_VISION_LLM,
-                "prompt": prompts.DEFAULT_MD_TABLE_PARSE_PROMPT,
-            }
-
-        if parse_images:
-            if image_args is None:
-                warnings.warn(
-                    "`parse_images` is set to `True`, but `image_args` is not specified, defaulting to `gpt-4o`."
-                )
-                image_args = {
-                    "parsing_algorithm": "llm",
-                    "llm": DEFAULT_VISION_LLM,
-                    "prompt": prompts.DEFAULT_IMAGE_PARSE_PROMPT,
-                }
-            else:
-                if image_args["parsing_algorithm"] != "llm":
-                    raise ValueError(
-                        "Image parsing is only supported with LLMs.",
-                        "Either change the `parsing_algorithm` to `llm` or set the `parse_images` to `False`.",
-                        f"Given args: {image_args}",
-                    )
-        else:
-            if image_args:
-                warnings.warn(
-                    "`parse_images` is set to `False`, but `image_args` is specified, skipping image parsing."
-                )
-                image_args = None
-
-        if processing_pipeline is None:
-            processing_pipeline = SimpleIngestionPipeline()
-        elif isinstance(processing_pipeline, str):
-            if processing_pipeline == "pathway_pdf_default":
-                processing_pipeline = SimpleIngestionPipeline()
-            elif processing_pipeline == "merge_same_page":
-                processing_pipeline = SamePageIngestionPipeline()
-            else:
-                raise ValueError(
-                    "Invalid `processing_pipeline` set. It must be either one of \
-                                 `'pathway_pdf_default'` or `'merge_same_page'`."
-                )
-
-        self.doc_parser = PyMuDocumentParser(
-            table_args=table_args,
-            image_args=image_args,
-            processing_pipeline=processing_pipeline,
-        )
-
-    async def __wrapped__(self, contents: bytes) -> list[tuple[str, dict]]:
-        import openparse
-        from pypdf import PdfReader
-
-        reader = PdfReader(stream=BytesIO(contents))
-        doc = openparse.Pdf(file=reader)
-
-        parsed_content = self.doc_parser.parse(doc)
-        nodes = list(parsed_content.nodes)
-
-        logger.info(
-            f"OpenParser completed parsing, total number of nodes: {len(nodes)}"
-        )
-
-        metadata: dict = {}
-        docs = [(node.model_dump()["text"], metadata) for node in nodes]
-
-        return docs
-
-    def __call__(self, contents: pw.ColumnExpression) -> pw.ColumnExpression:
-        """
-        Parse the given PDFs.
-
-        Args:
-            contents (ColumnExpression[bytes]): A column with PDFs to be parsed, passed as bytes.
-
-        Returns:
-            A column with a list of pairs for each query. Each pair is a text chunk and
-            metadata, which in case of `OpenParse` is an empty dictionary.
-        """
-        return super().__call__(contents)
 
 
 class ImageParser(pw.UDF):