+++ python/pathway/xpacks/llm/tests/test_parsers.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import os
 from pathlib import Path
 
 import nltk
@@ -12,12 +11,7 @@ from fpdf import FPDF
 
 import pathway as pw
 from pathway.tests.utils import assert_table_equality
-from pathway.xpacks.llm.parsers import (
-    OpenParse,
-    ParseUnstructured,
-    ParseUtf8,
-    PypdfParser,
-)
+from pathway.xpacks.llm.parsers import ParseUnstructured, ParseUtf8, PypdfParser
 
 for _ in range(10):
     try:
@@ -83,54 +77,6 @@ def _create_temp_pdf_with_text(text: str, path: Path) -> Path:
     pdf.output(pdf_path)
 
     return pdf_path
-
-
-@pytest.mark.environment_changes
-def test_parse_openparse(monkeypatch: pytest.MonkeyPatch, tmp_path: Path):
-    for tesseract_dir in [
-        "/usr/share/tesseract/tessdata/",
-        "/usr/share/tesseract-ocr/5/tessdata",
-    ]:
-        if os.path.exists(tesseract_dir):
-            monkeypatch.setenv(
-                "TESSDATA_PREFIX", tesseract_dir
-            )  # fix for tesseract ocr
-            break
-
-    if not os.getenv("TESSDATA_PREFIX"):
-        monkeypatch.setenv("TESSDATA_PREFIX", "/usr/share/tesseract/tessdata/")
-
-    txt = (
-        "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod"
-        "tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,"
-        "quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat."
-    )
-
-    pdf_path = _create_temp_pdf_with_text(txt, tmp_path)
-
-    parser = OpenParse(table_args={"parsing_algorithm": "pymupdf"})
-
-    with open(pdf_path, "rb") as pdf_file:
-        raw_pdf_data = pdf_file.read()
-
-    input_df = pd.DataFrame([dict(raw=raw_pdf_data)])
-
-    class Schema(pw.Schema):
-        raw: bytes
-
-    @pw.udf
-    def remove_line_breaks(t: str) -> str:
-        return t.replace("<br>", " ").replace("  ", " ")
-
-    input_table = pw.debug.table_from_pandas(input_df, schema=Schema)
-    result = input_table.select(ret=parser(pw.this.raw)[0][0])
-    # pdf creator adds line breaks
-    result = result.with_columns(ret=remove_line_breaks(pw.this.ret))
-
-    expected_df = pd.DataFrame([dict(ret=txt)])
-    expected_table = pw.debug.table_from_pandas(expected_df)
-
-    assert_table_equality(result, expected_table)
 
 
 def test_parse_pypdf(tmp_path: Path):